requirements.txt
arrow==0.14.7
canvasapi==0.10.0
certifi==2020.12.5
chardet==3.0.4
click==6.7
ics==0.7
idna==2.8
numpy==1.19.5
pandas==1.2.0
python-dateutil==2.8.1
python-dotenv==0.8.2
pytz==2020.5
requests==2.21.0
six==1.15.0
TatSu==5.5.0
urllib3==1.24.3
